refactor(manufacture): extract FlexiLotLoader
