+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/Internal/IFlexiLotLoader.cs
@@ -1,0 +1,10 @@
+using Anela.Heblo.Domain.Features.Catalog.Lots;
+
+namespace Anela.Heblo.Adapters.Flexi.Manufacture.Internal;
+
+internal interface IFlexiLotLoader
+{
+    Task<Dictionary<string, List<CatalogLot>>> LoadAvailableLotsAsync(
+        Dictionary<string, IngredientRequirement> ingredientRequirements,
+        CancellationToken cancellationToken);
+}
